Modificado y arreglado tema pagina cliente
